[Mod] 把pack_section放到segmenter.py复用
